fix(gameactions): moved to model
